+++ src/js/experimental/media-chrome-listbox.js
@@ -3,34 +3,38 @@ import { window, document } from '../utils/server-safe-globals.js';
 
 const template = document.createElement('template');
 
-template.innerHTML = `
+template.innerHTML = /*html*/`
 <style>
   :host ul {
+    font: var(--media-font,
+      var(--media-font-weight, normal)
+      var(--media-font-size, 1em) /
+      var(--media-text-content-height, var(--media-control-height, 24px))
+      var(--media-font-family, helvetica neue, segoe ui, roboto, arial, sans-serif));
+    color: var(--media-text-color, var(--media-primary-color, rgb(238 238 238)));
+    background: var(--media-listbox-background, var(--media-control-background, var(--media-secondary-color, rgb(20 20 30 / .8))));
     list-style: none;
     display: inline-flex;
     flex-direction: column;
     gap: 0.5em;
     margin: 0;
     padding: 0.5em;
-    background-color: var(--media-listbox-background, var(--media-control-background, rgba(10,10,15, .8)));
-    color: var(--media-text-color, white);
-    font-family: Arial, sans-serif;
   }
 
   ::slotted(media-chrome-listitem[tabindex="0"]:focus-visible),
   media-chrome-listitem[tabindex="0"]:focus-visible {
-    box-shadow: inset 0 0 0 2px rgba(27, 127, 204, 0.9);
+    box-shadow: inset 0 0 0 2px rgb(27 127 204 / .9);
     outline: 0;
   }
 
   ::slotted(media-chrome-listitem[aria-selected="true"]),
   media-chrome-listitem[aria-selected="true"] {
-    background-color: var(--media-listbox-selected-background, rgba(122,122,184, .8));
+    background-color: var(--media-listbox-selected-background, rgb(122 122 184 / .8));
   }
 
   ::slotted(media-chrome-listitem:hover),
   media-chrome-listitem:hover {
-    background-color: var(--media-listbox-hover-background, rgba(82,82,122, .8));
+    background-color: var(--media-listbox-hover-background, rgb(82 82 122 / .8));
     outline: var(--media-listbox-hover-outline, none);
   }
 </style>
@@ -47,6 +51,7 @@ class MediaChromeListbox extends window.HTMLElement {
   #clearKeysTimeout = null;
   #slot;
   #_assignedElements;
+  #metaPressed = false;
 
   static get observedAttributes() {
     return ['disabled', MediaStateReceiverAttributes.MEDIA_CONTROLLER];
@@ -55,21 +60,24 @@ class MediaChromeListbox extends window.HTMLElement {
   constructor(options = {}) {
     super();
 
-    const shadow = this.attachShadow({ mode: 'open' });
+    if (!this.shadowRoot) {
+      // Set up the Shadow DOM if not using Declarative Shadow DOM.
+      const shadow = this.attachShadow({ mode: 'open' });
 
-    const listboxHTML = template.content.cloneNode(true);
-    this.nativeEl = listboxHTML;
+      const listboxHTML = template.content.cloneNode(true);
+      this.nativeEl = listboxHTML;
 
-    let slotTemplate = options.slotTemplate;
+      let slotTemplate = options.slotTemplate;
 
-    if (!slotTemplate) {
-      slotTemplate = document.createElement('template');
-      slotTemplate.innerHTML = `<slot>${options.defaultContent || ''}</slot>`;
+      if (!slotTemplate) {
+        slotTemplate = document.createElement('template');
+        slotTemplate.innerHTML = `<slot>${options.defaultContent || ''}</slot>`;
+      }
+
+      this.nativeEl.appendChild(slotTemplate.content.cloneNode(true));
+
+      shadow.appendChild(listboxHTML);
     }
-
-    this.nativeEl.appendChild(slotTemplate.content.cloneNode(true));
-
-    shadow.appendChild(listboxHTML);
 
     this.#slot = this.shadowRoot.querySelector('slot');
 
@@ -126,12 +134,34 @@ class MediaChromeListbox extends window.HTMLElement {
     return this.#items.filter(el => el.getAttribute('aria-selected') === 'true');
   }
 
+  get value() {
+    return this.selectedOptions[0].value || this.selectedOptions[0].textContent;
+  }
+
+  set value(newValue) {
+    const item = this.#items.find(el => el.value === newValue || el.textContent === newValue);
+
+    if (!item) return;
+
+    this.#selectItem(item);
+  }
+
   focus() {
     this.selectedOptions[0]?.focus();
   }
 
   #clickListener = (e) => {
     this.handleClick(e);
+  }
+
+  #handleKeyListener(e) {
+    const { key } = e;
+
+    if (key === 'Enter' || key === ' ') {
+      this.handleSelection(e, this.hasAttribute('aria-multiselectable') && this.getAttribute('aria-multiselectable') === 'true');
+    } else {
+      this.handleMovement(e);
+    }
   }
 
   // NOTE: There are definitely some "false positive" cases with multi-key pressing,
@@ -144,20 +174,37 @@ class MediaChromeListbox extends window.HTMLElement {
       return;
     }
 
-    if (key === 'Enter' || key === ' ') {
-      this.handleSelection(e);
-    } else {
-      this.handleMovement(e);
+    if (key === 'Meta') {
+      this.#metaPressed = false;
+      return;
     }
+
+    this.#handleKeyListener(e);
   }
 
   #keydownListener = (e) => {
-    const { metaKey, altKey } = e;
-    if (metaKey || altKey) {
+    const { key, altKey } = e;
+
+    if (altKey) {
       this.removeEventListener('keyup', this.#keyupListener);
       return;
     }
-    e.preventDefault();
+
+    if (key === 'Meta') {
+      this.#metaPressed = true;
+      return;
+    }
+
+    // only prevent default on used keys
+    if (this.keysUsed.includes(key)) {
+      e.preventDefault();
+    }
+
+    if (this.#metaPressed && this.keysUsed.includes(key)) {
+      this.#handleKeyListener(e);
+      return;
+    }
+
     this.addEventListener('keyup', this.#keyupListener, {once: true});
   }
 
@@ -234,19 +281,27 @@ class MediaChromeListbox extends window.HTMLElement {
     return composedPath[index];
   }
 
-  handleSelection(e) {
+  handleSelection(e, toggle) {
     const item = this.#getItem(e);
 
     if (!item) return;
 
-    const selected = item.getAttribute('aria-selected') === 'true';
+    this.#selectItem(item, toggle);
+  }
 
-    if (this.getAttribute('aria-multiselectable') !== 'true') {
+  #selectItem(item, toggle) {
+    if (!this.hasAttribute('aria-multiselectable') || this.getAttribute('aria-multiselectable') !== 'true') {
       this.#assignedElements.forEach(el => el.setAttribute('aria-selected', 'false'));
     }
 
-    if (selected) {
-      item.setAttribute('aria-selected', 'false');
+    if (toggle) {
+      const selected = item.getAttribute('aria-selected') === 'true';
+
+      if (selected) {
+        item.setAttribute('aria-selected', 'false');
+      } else {
+        item.setAttribute('aria-selected', 'true');
+      }
     } else {
       item.setAttribute('aria-selected', 'true');
     }
@@ -308,7 +363,7 @@ class MediaChromeListbox extends window.HTMLElement {
     this.#items.forEach(el => el.setAttribute('tabindex', '-1'));
     item.setAttribute('tabindex', '0');
 
-    this.handleSelection(e);
+    this.handleSelection(e, this.hasAttribute('aria-multiselectable') && this.getAttribute('aria-multiselectable') === 'true');
   }
 
   #searchItem(key) {