+++ src/js/media-chrome-button.js
@@ -4,29 +4,25 @@ import { window, document } from './utils/server-safe-globals.js';
 
 const template = document.createElement('template');
 
-template.innerHTML = `
+template.innerHTML = /*html*/`
 <style>
   :host {
+    font: var(--media-font,
+      var(--media-font-weight, bold)
+      var(--media-font-size, 14px) /
+      var(--media-text-content-height, var(--media-control-height, 24px))
+      var(--media-font-family, helvetica neue, segoe ui, roboto, arial, sans-serif));
+    color: var(--media-text-color, var(--media-primary-color, rgb(238 238 238)));
+    background: var(--media-control-background, var(--media-secondary-color, rgb(20 20 30 / .7)));
+    padding: var(--media-control-padding, 10px);
     display: inline-flex;
     align-items: center;
     justify-content: center;
     vertical-align: middle;
     box-sizing: border-box;
-    background: var(--media-control-background, rgba(20,20,30, 0.7));
-
-    padding: var(--media-control-padding, 10px);
-
-    ${/* Vertically center any text */''}
-    font-size: 14px;
-    line-height: var(--media-text-content-height, var(--media-control-height, 24px));
-    font-weight: bold;
-    color: #fff;
-
-    transition: background 0.15s linear;
-
+    transition: background .15s linear;
     pointer-events: auto;
     cursor: pointer;
-    font-family: Arial, sans-serif;
   }
 
   ${/*
@@ -34,7 +30,7 @@ template.innerHTML = `
     https://drafts.csswg.org/selectors-4/#the-focus-visible-pseudo
   */''}
   :host(:focus-visible) {
-    box-shadow: inset 0 0 0 2px rgba(27, 127, 204, 0.9);
+    box-shadow: inset 0 0 0 2px rgb(27 127 204 / .9);
     outline: 0;
   }
   ${/*
@@ -46,7 +42,7 @@ template.innerHTML = `
   }
 
   :host(:hover) {
-    background: var(--media-control-hover-background, rgba(50,50,70, 0.7));
+    background: var(--media-control-hover-background, rgba(50 50 70 / .7));
   }
 
   svg, img, ::slotted(svg), ::slotted(img) {
@@ -54,7 +50,7 @@ template.innerHTML = `
     height: var(--media-button-icon-height, var(--media-control-height, 24px));
     transform: var(--media-button-icon-transform);
     transition: var(--media-button-icon-transition);
-    fill: var(--media-icon-color, #eee);
+    fill: var(--media-icon-color, var(--media-primary-color, rgb(238 238 238)));
     vertical-align: middle;
     max-width: 100%;
     max-height: 100%;
@@ -77,22 +73,25 @@ class MediaChromeButton extends window.HTMLElement {
   constructor(options = {}) {
     super();
 
-    this.attachShadow({ mode: 'open' });
+    if (!this.shadowRoot) {
+      // Set up the Shadow DOM if not using Declarative Shadow DOM.
+      this.attachShadow({ mode: 'open' });
 
-    const buttonHTML = template.content.cloneNode(true);
-    this.nativeEl = buttonHTML;
+      const buttonHTML = template.content.cloneNode(true);
+      this.nativeEl = buttonHTML;
 
-    // Slots
-    let slotTemplate = options.slotTemplate;
+      // Slots
+      let slotTemplate = options.slotTemplate;
 
-    if (!slotTemplate) {
-      slotTemplate = document.createElement('template');
-      slotTemplate.innerHTML = `<slot>${options.defaultContent || ''}</slot>`;
+      if (!slotTemplate) {
+        slotTemplate = document.createElement('template');
+        slotTemplate.innerHTML = `<slot>${options.defaultContent || ''}</slot>`;
+      }
+
+      this.nativeEl.appendChild(slotTemplate.content.cloneNode(true));
+
+      this.shadowRoot.appendChild(buttonHTML);
     }
-
-    this.nativeEl.appendChild(slotTemplate.content.cloneNode(true));
-
-    this.shadowRoot.appendChild(buttonHTML);
 
     const { style } = getOrInsertCSSRule(this.shadowRoot, ':host');
     style.setProperty('display', `var(--media-control-display, var(--${this.localName}-display, inline-flex))`);
@@ -113,7 +112,9 @@ class MediaChromeButton extends window.HTMLElement {
       return;
     }
 
-    this.handleClick(e);
+    if (!this.preventClick) {
+      this.handleClick(e);
+    }
   }
 
   #keydownListener = (e) => {
@@ -133,6 +134,7 @@ class MediaChromeButton extends window.HTMLElement {
 
   disable() {
     this.removeEventListener('click', this.#clickListener);
+    this.removeEventListener('keydown', this.#keydownListener);
     this.removeEventListener('keyup', this.#keyupListener);
     this.removeAttribute('tabindex');
   }