+++ src/js/media-container.js
@@ -188,9 +188,11 @@ class MediaContainer extends window.HTMLElement {
   constructor() {
     super();
 
-    // Set up the Shadow DOM
-    this.attachShadow({ mode: 'open' });
-    this.shadowRoot.appendChild(template.content.cloneNode(true));
+    if (!this.shadowRoot) {
+      // Set up the Shadow DOM if not using Declarative Shadow DOM.
+      this.attachShadow({ mode: 'open' });
+      this.shadowRoot.appendChild(template.content.cloneNode(true));
+    }
 
     // Watch for child adds/removes and update the media element if necessary
     const mutationCallback = (mutationsList) => {
@@ -374,7 +376,10 @@ class MediaContainer extends window.HTMLElement {
 
     const setInactive = () => {
       if (this.autohide < 0) return;
-      this.setAttribute('user-inactive', 'user-inactive');
+      if (this.hasAttribute('user-inactive')) return;
+
+      this.setAttribute('user-inactive', '');
+
       const evt = new window.CustomEvent(
         MediaStateChangeEvents.USER_INACTIVE,
         { composed: true, bubbles: true, detail: true }
@@ -383,7 +388,10 @@ class MediaContainer extends window.HTMLElement {
     };
 
     const setActive = () => {
+      if (!this.hasAttribute('user-inactive')) return;
+
       this.removeAttribute('user-inactive');
+
       const evt = new window.CustomEvent(
         MediaStateChangeEvents.USER_INACTIVE,
         { composed: true, bubbles: true, detail: false }