+++ src/js/media-gesture-receiver.js
@@ -30,21 +30,24 @@ class MediaGestureReceiver extends window.HTMLElement {
   constructor(options = {}) {
     super();
 
-    const shadow = this.attachShadow({ mode: 'open' });
+    if (!this.shadowRoot) {
+      // Set up the Shadow DOM if not using Declarative Shadow DOM.
+      const shadow = this.attachShadow({ mode: 'open' });
 
-    const buttonHTML = template.content.cloneNode(true);
-    this.nativeEl = buttonHTML;
+      const buttonHTML = template.content.cloneNode(true);
+      this.nativeEl = buttonHTML;
 
-    // Slots
-    let slotTemplate = options.slotTemplate;
+      // Slots
+      let slotTemplate = options.slotTemplate;
 
-    if (!slotTemplate) {
-      slotTemplate = document.createElement('template');
-      slotTemplate.innerHTML = `<slot>${options.defaultContent || ''}</slot>`;
+      if (!slotTemplate) {
+        slotTemplate = document.createElement('template');
+        slotTemplate.innerHTML = `<slot>${options.defaultContent || ''}</slot>`;
+      }
+
+      this.nativeEl.appendChild(slotTemplate.content.cloneNode(true));
+      shadow.appendChild(buttonHTML);
     }
-
-    this.nativeEl.appendChild(slotTemplate.content.cloneNode(true));
-    shadow.appendChild(buttonHTML);
   }
 
   attributeChangedCallback(attrName, oldValue, newValue) {