+++ src/js/media-theme-element.js
@@ -39,10 +39,18 @@ export class MediaThemeElement extends window.HTMLElement {
   renderer;
   #template;
   #prevTemplate;
+  #prevTemplateId;
 
   constructor() {
     super();
-    this.renderRoot = this.attachShadow({ mode: 'open' });
+
+    if (this.shadowRoot) {
+      this.renderRoot = this.shadowRoot;
+    } else {
+      // Set up the Shadow DOM if not using Declarative Shadow DOM.
+      this.renderRoot = this.attachShadow({ mode: 'open' });
+      this.createRenderer();
+    }
 
     const observer = new MutationObserver((mutationList) => {
       if (mutationList.some((mutation) => {
@@ -75,8 +83,6 @@ export class MediaThemeElement extends window.HTMLElement {
       subtree: true,
     });
 
-    this.createRenderer();
-
     // In case the template prop was set before custom element upgrade.
     // https://web.dev/custom-elements-best-practices/#make-properties-lazy
     this.#upgradeProperty('template');
@@ -103,6 +109,7 @@ export class MediaThemeElement extends window.HTMLElement {
   }
 
   set template(element) {
+    this.#prevTemplateId = null;
     this.#template = element;
     this.createRenderer();
   }
@@ -140,30 +147,46 @@ export class MediaThemeElement extends window.HTMLElement {
     return props;
   }
 
-  async attributeChangedCallback(attrName, oldValue, newValue) {
+  attributeChangedCallback(attrName, oldValue, newValue) {
     if (attrName === 'template' && oldValue != newValue) {
+      this.#updateTemplate();
+    }
+  }
 
-      if (newValue) {
-        // First try to get a template element by id
-        // @ts-ignore
-        const template = this.getRootNode()?.getElementById(newValue);
-        if (template) {
-          this.#template = template;
+  connectedCallback() {
+    this.#updateTemplate();
+  }
 
-        } else if (isValidUrl(newValue)) {
-          // Next try to fetch a HTML file if it looks like a valid URL.
-          try {
-            const data = await request(newValue);
-            const template = document.createElement('template');
-            template.innerHTML = data;
-            this.#template = template;
-          } catch {
-            console.warn(`"${newValue}" is not a valid template element ID or a valid URL to a template file.`);
-          }
-        }
-      }
+  #updateTemplate() {
+    const templateId = this.getAttribute('template');
+    if (!templateId || templateId === this.#prevTemplateId) return;
 
+    // First try to get a template element by id
+    const rootNode = /** @type HTMLDocument | ShadowRoot */ (this.getRootNode());
+    const template = rootNode?.getElementById?.(templateId);
+
+    if (template) {
+      // Only save prevTemplateId if a template was found.
+      this.#prevTemplateId = templateId;
+      this.#template = template;
       this.createRenderer();
+      return;
+    }
+
+    if (isValidUrl(templateId)) {
+      // Save prevTemplateId on valid URL before async fetch to prevent duplicate fetch.
+      this.#prevTemplateId = templateId;
+
+      // Next try to fetch a HTML file if it looks like a valid URL.
+      request(templateId)
+        .then((data) => {
+          const template = document.createElement('template');
+          template.innerHTML = data;
+
+          this.#template = template;
+          this.createRenderer();
+        })
+        .catch(console.error);
     }
   }
 
@@ -192,6 +215,9 @@ export class MediaThemeElement extends window.HTMLElement {
 }
 
 function isValidUrl(url) {
+  // Valid URL e.g. /absolute, ./relative, http://, https://
+  if (!/^(\/|\.\/|https?:\/\/)/.test(url)) return false;
+
   // Add base if url doesn't start with http:// or https://
   const base = /^https?:\/\//.test(url) ? undefined : location.origin;
   try {