+++ src/js/themes/youtube.js
@@ -14,18 +14,16 @@ const template = document.createElement('template');
 template.innerHTML = `
 <style>
   :host {
-    --primary-color: #eee;
-    --secondary-color: transparent;
+    --media-secondary-color: transparent;
+    --media-primary-color: #fff;
   }
 
   media-controller {
     font-size: 13px;
     font-family: Roboto, Arial, sans-serif;
     -webkit-font-smoothing: antialiased;
-    color: var(--primary-color);
 
-    --media-control-background: var(--secondary-color);
-    --media-control-hover-background: var(--secondary-color);
+    --media-control-hover-background: var(--media-secondary-color);
     --media-range-track-height: 3px;
     --media-range-thumb-height: 13px;
     --media-range-thumb-width: 13px;
@@ -34,12 +32,6 @@ template.innerHTML = `
     --media-preview-thumbnail-border-radius: 2px;
     --media-preview-time-margin: 10px 0;
 
-    --media-icon-color: var(--primary-color);
-  }
-
-  *:hover {
-    color: #fff;
-    --media-icon-color: #fff;
   }
 
   media-control-bar {
@@ -82,8 +74,6 @@ template.innerHTML = `
     padding-left: 0px;
 
     --media-range-track-background: rgba(255,255,255,.2);
-    --media-range-bar-color: #fff;
-    --media-range-thumb-background: #fff;
   }
 
   .control-spacer {